videodeepsearch/pyproject.toml
[project]
name = "videodeepsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix>=12.12.0",
    "asyncpg>=0.30.0",
    "av>=16.0.1",
    "fastapi[standard]>=0.121.0",
    "llama-index>=0.14.7",
    "llama-index-callbacks-arize-phoenix>=0.6.1",
    "llama-index-core>=0.14.7",
    "llama-index-llms-google-genai>=0.7.2",
    "llama-index-workflows>=2.11.1",
    "loguru>=0.7.3",
    "minio>=7.2.18",
    "opencv-python>=4.12.0.88",
    "pydantic-settings>=2.11.0",
    "pymilvus>=2.6.3",
]
